Large re-org in preparation for EEPROM and comms paging work

+++ table.h
@@ -4,8 +4,8 @@ This file is used for everything related to maps/tables including their definiti
 #include <Arduino.h>
 struct table {
   //All tables must be the same size for simplicity
-  const static int xSize = 8;
-  const static int ySize = 8;
+  const static byte xSize = 8;
+  const static byte ySize = 8;
   
   byte values[ySize][xSize];
   int axisX[xSize];
